+++ src/types.ts
@@ -10,11 +10,12 @@ interface SystemApi {
     actorRollSkill: (actor, skillId: string) => Promise<Roll>;
     actorRollAbility: (actor, abilityId: string) => Promise<Roll>;
     actorRollTool?: (actor,item) => Promise<Roll>;
+    actorCurrenciesAdd?: (actor, currencies: Currencies) => Promise<void>; //deprecated
     actorCurrenciesGet?: (actor) => Currencies;
-    actorCurrenciesAdd?: (actor, currencies: Currencies) => Promise<void>; //may throw Error
+    actorCurrenciesStore?: (actor, currencies: Currencies) => Promise<void>;
     actorSheetAddTab:(sheet, html, actor, tabData: { id: string, label: string, html: string }, tabBody: string) => void;
     componentIsSame?:(a: ComponentData,b: ComponentData)=>boolean,
-    componentFromEntity?:(entity)=>Component,
+    componentFromEntity?:(entity,hasJsonData?:boolean)=>Component,
     componentDefaultData?: ComponentData,
     itemQuantityAttribute:string,
     itemPriceAttribute:string,
@@ -31,13 +32,14 @@ interface System extends SystemApi {
     actorCurrenciesGet: (actor) => Currencies;
     actorCurrenciesAdd: (actor, currencies: Currencies) => Promise<void>; //may throw Error
     actorCurrenciesCanAdd:(actor, currencies: Currencies)=>boolean;
-    actorComponentListAdd:(actor,componentList: Component[])=>Promise<void>,
+    actorComponentListAdd:(actor,componentList: Component[])=>Promise<ItemChange>,
     uuidToDocument: (string)=>Promise<foundry.abstract.Document<any, any>>
     componentCreate:(data) => Component
     componentDefaultData: ComponentData,
-    componentFromEntity:(entity)=>Component,
+    componentFromEntity:(entity,hasJsonData?:boolean)=>Component,
     componentIsSame:(a: ComponentData,b: ComponentData)=>boolean,
-    objectAttributeGet:(obj:any, attribute:string)=>any,
+    currenciesSum:(source: Currencies, add: Currencies, doExchange:boolean)=>Currencies
+    objectAttributeGet:(obj:any, attribute:string, fallback?:any)=>any,
     objectAttributeSet:(obj:any, attribute:string, value)=>void,
     itemListComponentFind:(itemList,component: ComponentData)=>{components:Component[],quantity:number},
     uiDialogSelect:(data: SelectData)=>Promise<string>
@@ -75,8 +77,8 @@ interface ComponentData {
     img: string;
     quantity: number;
     itemType?: string;      //if it is of type item there is an itemType
+    jsonData?: string;      //to store a component completly
     [key: string]: unknown; //this is system dependent information! do not relay on it. It is only needed for internal behavior e.g. isSame.
-
 }
 
 /**
@@ -102,4 +104,10 @@ interface CurrencyConfig {
     factor: number, //factor how often the lowest currency fits into this currency
     uuid?: string, //system dependent if this is an item
     component?: Component //will get automatically attached when an uuid is given
+}
+interface ItemChange {
+    create: any[]
+    update: any[],
+    merge: string[],
+    delete: ComponentData[]
 }